Implementing cache eviction based on a ttl value. (#67)

+++ stundb/net/core/src/main/java/com/stundb/net/core/models/requests/SetRequest.java
@@ -1,4 +1,4 @@
 package com.stundb.net.core.models.requests;
 
-public record SetRequest(String key, Object value, Integer ttl) {
+public record SetRequest(String key, Object value, Long ttl) {
 }
